backend/requirements.txt
Flask
Flask-Cors
Pandas
sklearn.model_selection
sklearn.linear_model
sklearn.metrics